fix: handle ftell() errors + clean up file-offset sign conversions
The -Wsign-conversion sweep (commit aa66b35) deliberately deferred the
sign-conversion warnings. This commit addresses the subset that can mask
real errors: long/off_t/ssize_t -> size_t conversions in file-offset
code, where a negative value means error.

safetensors.c (real bug fix):
- 5 ftell()-derived sizes were fed straight into malloc()/fread() with no
  error check. ftell() returns -1 on failure; that -1 (as size_t = huge)
  would hit malloc/fread. Now each site checks `ftell() < 0`:
  required reads (load_config, load_tokenizer vocab) return -1; optional
  reads (rope params, tokenizer cfg) clamp to empty. The size is then
  converted to size_t for downstream use.

model.c:
- Cast st.st_size -> size_t in the two mmap() calls (regular files, >= 0).
- Cast pointer-difference -> size_t for the mlock/munlock range-size
  computations (3 sites: the ADD_RANGE macro + the two merged-range loops;
  end >= start by construction) and the GGUF reader r.pos = pp - r.data.

server.c:
- KV-cache slot size accounting mixed off_t with size_t (sizeof/strides),
  dropping to unsigned mid-expression. Cast the sizeof products and the
  stride sum to off_t so the offset arithmetic stays in off_t (kv_end's
  type) -- all operands are positive counts.

Result: 0 off_t/long/ssize_t->size_t sign-conversions remain; total
sign-conversion drops 471 -> 438. No new warnings; build links.
The remaining ~438 are the safe int<->size_t dimension math (per
aa66b35's stance) and are intentionally left.

+++ picolm/model.c
@@ -657,7 +657,7 @@ int model_list_tensors(const char *path) {
 
     { struct stat st;
       if (fstat(fd, &st) < 0) goto out;
-      addr = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
+      addr = mmap(NULL, (size_t)st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
       if (addr == MAP_FAILED) { addr = NULL; goto out; }
       fsize = (size_t)st.st_size; }
 #endif
@@ -788,7 +788,7 @@ int model_list_kv(const char *path) {
     if (fd < 0) { perror("open"); return -1; }
     { struct stat st;
       if (fstat(fd, &st) < 0) goto out;
-      addr = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
+      addr = mmap(NULL, (size_t)st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
       if (addr == MAP_FAILED) { addr = NULL; goto out; }
       fsize = (size_t)st.st_size; }
 #endif
@@ -1135,7 +1135,7 @@ static int parse_gguf(model_t *m, int max_seq_len) {
                   uint64_t nn = arr_len;
                   for (uint64_t ii = 0; ii < nn; ii++) {
                       uint64_t sl; memcpy(&sl, pp, 8); sl = GGUF_LE64(sl); pp += 8 + sl; }
-                  r.pos = pp - r.data; }
+                  r.pos = (size_t)(pp - r.data); }
             }
         } else if (str_eq(key, "tokenizer.ggml.token_type")) {
             if (vtype != GGUF_META_ARRAY) {
@@ -6003,7 +6003,7 @@ int model_lock_layers(model_t *m, size_t mem_bytes) {
         if (ptr && _sz > 0 && n_ranges < NR_MAX) { \
             ranges_start[n_ranges] = page_align_down(ptr); \
             ranges_end[n_ranges] = page_align_up((const uint8_t *)ptr + _sz); \
-            ranges_size[n_ranges] = (const uint8_t *)ranges_end[n_ranges] - (const uint8_t *)ranges_start[n_ranges]; \
+            ranges_size[n_ranges] = (size_t)((const uint8_t *)ranges_end[n_ranges] - (const uint8_t *)ranges_start[n_ranges]); \
             n_ranges++; \
         } \
     } while(0)
@@ -6118,7 +6118,7 @@ int model_lock_layers(model_t *m, size_t mem_bytes) {
     /* Call mlock on each merged range */
     size_t total_locked = 0;
     for (int i = 0; i < merged; i++) {
-        size_t sz = (const uint8_t *)ranges_end[i] - (const uint8_t *)ranges_start[i];
+        size_t sz = (size_t)((const uint8_t *)ranges_end[i] - (const uint8_t *)ranges_start[i]);
 #ifdef _WIN32
         if (VirtualLock((void *)(const void *)ranges_start[i], sz) == 0) {
             fprintf(stderr, "Lock: VirtualLock failed (error %lu)\n", (unsigned long)GetLastError());
@@ -6221,7 +6221,7 @@ int model_unlock_layers(model_t *m) {
 
     size_t total_unlocked = 0;
     for (int i = 0; i < merged; i++) {
-        size_t sz = (const uint8_t *)ranges_end[i] - (const uint8_t *)ranges_start[i];
+        size_t sz = (size_t)((const uint8_t *)ranges_end[i] - (const uint8_t *)ranges_start[i]);
 #ifdef _WIN32
         VirtualUnlock((void *)(const void *)ranges_start[i], sz);
 #else
